requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
google-generativeai>=0.7.0
langgraph==0.2.16
langchain==0.2.16
langchain-google-genai==1.0.10
langchain-core==0.2.38
pydantic==2.7.4
uvicorn==0.30.0
gunicorn==21.2.0
requests==2.31.0
typing-extensions==4.12.2 